requirements.txt
numpy
pandas
scikit-learn
pyyaml
rosbags
opencv-python
scipy